Try to remove more background pressure signal.
There is some Δρ over the slope near the bottom before the eddy
gets there. I think this is associated with the bottom boundary
condition. To remove it, I subtract the density field at the
easternmost edge of the domain at the first instant. This will have
to be updated when I do the calculation for the entire run duration.