Updates to contour plots in runs.plot_fluxes.
1. Overlay idealized velocity profile.
2. Estimate the streamer flux numerically from actual output and
compare against my idealized estimate. I think the difference is
due to the streamer having different vertical scales in deeper water
when compared to the idealized version.

Physically, this is because the streamer is next to a bit of the
eddy that is about to filament. So there is a kink in the velocity
field that has scales smaller than the actual eddy.