Updates to maxflux parameterization.
Currently, this is my best yet for y/R = 0.5. This has happened
because of tweaks to calc_maxflux, flux_tindices and using
Lz(maxloc) for vertical scale. I think the eddy.V and eddy.Ls
time series are just a little to spiky at 'maxloc' and so, are
creating extra scatter.