Fix for sloping shelves.
Let's assume that SSH at (x/L) = -1 is 0. Then, the outflow of scale
L_β should "see" the SSH field between (x/L) = -1 and (x/L) = -1+L_β.
I then get rid of the (x/L) = -1 requirement for simplicity and so,
the mask is (x/L) < -1 + L_β.

Set default L_β to 1 so that for very gentle or flat shelves, the mask is
(x/L) < 0 as usual.